+++ hserv/utilities/uFile.php
@@ -627,7 +627,7 @@
     }    
     
     function fileDelete( $filename ){
-        if(file_exists($filename)){
+        if(!empty($filename) && file_exists($filename)){
             unlink($filename);
         }
     }
